+++ src/main/java/com/kusher/kusher_in_korea/tabling/dto/request/CreateReservationDto.java
@@ -10,6 +10,6 @@ public class CreateReservationDto { // 예약 추가 요청
     private Long restaurantId; // 식당(식당번호)
     private String reservationDate; // 예약날짜
     private String reservationTime; // 예약시간
-    private Long numberOfPeople; // 예약인원
+    private int numberOfPeople; // 예약인원
     private String status; // 예약상태
 }
